backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.11.7
motor==3.5.0
pymongo==4.8.0
python-dotenv==1.1.1
python-multipart==0.0.9
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
Pillow==10.4.0
stripe